refactor(rpc): separate basic rpc by group
