+++ packages/ckb-sdk-rpc/src/Base/pool.ts
@@ -1,0 +1,18 @@
+import paramsFmts from '../paramsFormatter'
+import resultFmts from '../resultFormatter'
+
+export default {
+  sendTransaction: {
+    method: 'send_transaction',
+    paramsFormatters: [paramsFmts.toRawTransaction, paramsFmts.toOutputsValidator],
+    resultFormatters: resultFmts.toHash,
+  },
+
+  txPoolInfo: {
+    method: 'tx_pool_info',
+    paramsFormatters: [],
+    resultFormatters: resultFmts.toTxPoolInfo,
+  },
+
+  // TODO: clear_tx_pool
+}
